chore: use scandata datasets

+++ eagle/train/eagle3/modeling_llama_kv.py
@@ -889,7 +889,7 @@ class LlamaModel(LlamaPreTrainedModel):
       combined_attention_mask = _make_causal_mask(
         input_shape,
         # inputs_embeds.dtype,
-        torch.float32,  # [MODIFIED] force to cast to float32
+        torch.float16,  # [MODIFIED] force to cast to float32
         device=inputs_embeds.device,
         past_key_values_length=past_key_values_length,
       )
